Source:
PeekBits: Commented out the bit-by-bit version and wrote a replacement
that goes byte by byte. Should be up to 8 times faster.

BitsRemaining: Tweaked it to be more sane, and to handle > 8 bits cases.

Power2Mask: Rewrote it to do it all in one step, should save a few
bytes of memory, and be a tad faster.

I still haven't cleaned out the old versions of PeekBits, incase I need
something, when I do It should make BitIO about 300 lines shorter.

PerfectSubtract: I'll probably rename it because damn that's a terrible
name, but the point is to never subtract something from a smaller
number, subtraction is position independent (in most cases, but not
all) so it's basically the default.

Log: Did a TINY bit of work to make it better, I may end up removing
the ErrorStatus stuff afterall.