Unconst'd the string and CommandLineInterface pointers
Removed FindCLIArgument, and added GetCLISwitchNumFromFlag.

in CommandLineInterface.c, I rearranged the structs to put any arrays
as far away from variables as possibly, hopefully that'll increase the
cache locality, and I renamed Meta* to Child*, and fixed any misuse of
Arguments ot Switches in the wrong type.

Made the DisplayProgramBanner more modular for when/if there's missing
information, and added preliminary support to ParseCommandLineARguments
to extract a single trailing child argument, I should rewrite it to
support more than one trailing argument, but that's not gonna happen in
this commit.